delivery request implementation

+++ src/main/java/com/algaworks/algalog/domain/service/CrudClienteService.java
@@ -1,8 +1,14 @@
 package com.algaworks.algalog.domain.service;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import com.algaworks.algalog.domain.exception.DomainException;
 import com.algaworks.algalog.domain.model.Cliente;
@@ -13,6 +19,11 @@ public class CrudClienteService {
 
 	@Autowired
 	private ClienteRepository clienteRepository;
+	
+	public Cliente buscar(Long clienteId) {
+		return clienteRepository.findById(clienteId)
+				.orElseThrow(() -> new DomainException("Cliente não encontrado"));
+	}
 	
 	@Transactional //deve ser executado dentro de uma transação
 	public Cliente salvar(Cliente cliente) {
@@ -28,8 +39,52 @@ public class CrudClienteService {
 	}
 	
 	@Transactional
-	public void excluir(Long clienteId) {
+	public ResponseEntity<Void> excluir(Long clienteId) {
+		if (!clienteRepository.existsById(clienteId)) {
+			return ResponseEntity.notFound().build();		
+		}
 		clienteRepository.deleteById(clienteId);		
+		return ResponseEntity.noContent().build();
+	}
+	
+	
+	@Transactional
+	public List<Cliente> listarAll(){
+		return clienteRepository.findAll();
+	}
+	
+	@Transactional
+	public List<Cliente> listarByNome(String nome){
+		return clienteRepository.findByNomeContaining(nome);
+	}
+	
+	@Transactional
+	public ResponseEntity<Cliente> listarById(Long id){
+		
+		//Optional é um container que nele pode ter ou nao algum valor
+		Optional<Cliente> cliente = clienteRepository.findById(id);
+		
+		//se dentro do Optional Cliente tem algo
+		if(cliente.isPresent()) {
+			//retorna uma resposta ok (200) e devolve o valor de cliente
+			return ResponseEntity.ok(cliente.get());		
+		}
+		else {
+			//caso o optional nao possuir algum valor dentro dele, retorna notFound (404)
+			return ResponseEntity.notFound().build();
+		}
+
+	}
+		
+	@Transactional
+	public ResponseEntity<Cliente> atualizarCliente(@PathVariable Long id, @RequestBody Cliente cliente){
+		if(!clienteRepository.existsById(id)) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		cliente.setId(id);
+		clienteRepository.save(cliente);
+		return ResponseEntity.ok(cliente);
 	}
 	
 }
